requirements.txt
blinker==1.9.0
boto3==1.39.2
botocore==1.39.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
imap-tools==1.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.3.1
pandas==2.2.2
pymongo
PyPDF2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests
s3transfer==0.13.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
